pyproject.toml
[tool.poetry]
name = "dfs"
version = "0.0.0"  # set via gitlab-ci
description = "lightweight pandas.DataFrame schema"
authors = ["Philipp <philippk@zillowgroup.com>"]
readme = "README.md"

[tool.poetry.scripts]
dfs = "dfs.cli:app"

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
pandas = "^1.2.4"
sqlalchemy = {version = "1.*", optional = true}
pandera = {version = "^0.6", optional = true}
pydantic = "^1.9.1"
typer = {version = "^0.4.1", optional = true}
PyYAML = {version = "^6.0", optional = true}


[tool.poetry.dev-dependencies]
hypothesis = "6.13.12"
pytest = "^5.4.1"
typer = "^0.4.1"
pre-commit = "^2.19.0"
jsonschema2md = "^0.4.0"

[tool.poetry.extras]
sqlalchemy = ["sqlalchemy"]
# pandera = ["pandera"]
cli = ["typer"]
yaml = ["PyYAML"]
all = ["sqlalchemy", "typer", "PyYAML"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
